weed_management/requirements.txt
Flask
tensorflow  # Or the version you're using for your CNN model
pillow  # For image processing
numpy
